requirements.txt
absl-py==0.2.0
astor==0.6.2
backcall==0.1.0
bleach==1.5.0
blessings==1.6.1
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
gast==0.2.0
gpustat==0.4.1
grpcio==1.11.0
html5lib==0.9999999
idna==2.6
imageio==2.3.0
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.3
nvidia-ml-py3==7.352.0
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==5.4.5
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
tqdm==4.23.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
